% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_fwf.R
\docType{package}
\name{fwf-package}
\alias{fwf-package}
\title{fwf: Fixed Width Format Tables}
\description{
\code{fwf} is a tiny package to read and write fixed width format tables.

\href{https://github.com/FinnishCancerRegistry/fwf/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/FinnishCancerRegistry/fwf/actions/workflows/R-CMD-check.yaml/badge.svg}}}
}
\section{Example}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Remember, data types other than character, integer, and double may not
# be properly handled (because meta-information about datatype is lost when
# writing into a text file). See ?data.table::fread for more control.
# over reading behaviour.
exp <- data.frame(
  col1 = c("abcdefghijklmn", "abc"),
  col2 = c(1e6L, 1L),
  col3 = as.Date(c("2000-01-01", "2001-01-01"))
)
tf <- tempfile(pattern = "fwf_file_", fileext = ".txt")
fwf::fwf_write(exp, tf)
obs <- fwf::fwf_read(
  path = tf,
  fread_arg_list = list(colClasses = c(col3 = "Date"))
)
stopifnot(
  all.equal(exp, obs, check.attributes = FALSE),
  identical(lapply(exp, class), lapply(obs, class))
)
}\if{html}{\out{</div>}}
}

\section{Installation}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github("FinnishCancerRegistry/fwf")
}\if{html}{\out{</div>}}
}

\section{News}{

\subsection{News for version 1.0.1}{
\subsection{fwf}{

Improve docpage ?fwf::fwf.
}

}

\subsection{News for version 1.0.0}{
\subsection{fwf}{

Bump to version 1.0.0.
}

}

\subsection{News for version 0.2.0}{
\subsection{fwf::fwf_write}{

Improved \code{fwf_write} internal preprocessing steps.
}

}

\subsection{News for version 0.1.0}{
\subsection{fwf::fwf_read}{

Fun \code{fwf_read} included in first version of this package.
}

\subsection{fwf::fwf_write}{

Fun \code{fwf_write} included in first version of this package.
}

}
}

\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.miettinen@cancer.fi} (\href{https://orcid.org/0000-0001-8624-6754}{ORCID})

}
\keyword{internal}
