% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwf.R, R/package_fwf.R
\docType{package}
\name{fwf}
\alias{fwf}
\alias{fwf_write}
\alias{fwf_read}
\title{Fixed Width Format}
\usage{
fwf_write(x, path, widths = NULL, fwrite_arg_list = NULL)

fwf_read(path, fread_arg_list = NULL)
}
\arguments{
\item{x}{\verb{[data.frame]} (no default)

\code{data.frame} to write to disk.}

\item{path}{\verb{[character]} (no default)
\itemize{
\item \code{fwf_write}: Path to write to.
\item \code{fwf_read}: Path to read from.
}}

\item{widths}{\verb{[NULL, integer]} (default \code{NULL})
\itemize{
\item \code{NULL}: the width of each field will be \code{max(nchar(x[[i]]))} for all
\verb{i in 1:ncol(x)}
\item \code{integer}: use these widths; must be of length \code{ncol(x)}
}}

\item{fwrite_arg_list}{\verb{[NULL]} (default \code{NULL})
\itemize{
\item \code{NULL}: No additional arguments passed to \link[data.table:fwrite]{data.table::fwrite}.
\item \code{list}: Pass these additional arguments passed to \link[data.table:fwrite]{data.table::fwrite}.
}

\link[data.table:fwrite]{data.table::fwrite} args \code{x} and \code{file} always set to args \code{x} and \code{path}
of this function.}

\item{fread_arg_list}{\verb{[NULL, list]} (default \code{NULL})
\itemize{
\item \code{NULL}: No additional arguments passed to \link[data.table:fread]{data.table::fread}.
\item \code{list}: Pass these additional arguments passed to \link[data.table:fread]{data.table::fread}.
}

\link[data.table:fread]{data.table::fread} arg \code{file} is always set to arg \code{path} of this function.}
}
\description{
Read and write fixed-width format files.

\code{fwf} is a tiny package to read and write fixed width format tables.
}
\section{News}{

\subsection{News for version 0.1.0}{
\subsection{Object fwf::fwf_read news:}{

Fun \code{fwf_read} included in first version of this package.
}

\subsection{Object fwf::fwf_write news:}{

Fun \code{fwf_write} included in first version of this package.
}

}
}

